requirements.txt
cloudformation-cli-python-lib==2.0.0
python-twitter>=3.5
